docker/backup_postgres/app/requirements.txt
fastapi[all]
uvicorn[standard]
psycopg2-binary
SQLAlchemy
python-dotenv
pydantic-settings
alembic # For database migrations, good practice
